% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model.interface.MCMCglmm.r
\name{check.supported.family}
\alias{check.supported.family}
\title{(Internal) Check if the specified family of MCMCglmm is supported.}
\usage{
check.supported.family(family)
}
\arguments{
\item{family}{a character vector of length one specifing}
}
\description{
(Internal) Check if the specified family of MCMCglmm is supported.
}
