% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.r
\name{convert.family}
\alias{convert.family}
\title{(Internal) Format and check consistency of family.}
\usage{
convert.family(family, type = c("family", "character"))
}
\arguments{
\item{family}{a family function, character of family name, symbol, or family object.}

\item{type}{a character to specify the type of the value this function returns.
     If "family", this function returns 'family' object. If "character" is
     specified, this function returns a character string denoting the name
     of the family.

     This internal function is called by \code{link{detect.model.type}} and
     \code{\link{predict.glmmML}} functions.}
}
\description{
(Internal) Format and check consistency of family.
}
