% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.r
\name{extract.family.from.call}
\alias{extract.family.from.call}
\title{(Internal) Get family from function call.}
\usage{
extract.family.from.call(x, envir = parent.frame())
}
\arguments{
\item{x}{a call of a function.}

\item{envir}{an environment where \code{x} is evaluated.}
}
\value{
returns family in the call if available else NULL.
}
\description{
(Internal) Get family from function call.
}
