% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.r
\name{find.original.name}
\alias{find.original.name}
\title{(Internal) Find original name of the function.}
\usage{
find.original.name(fun)
}
\arguments{
\item{fun}{a function}
}
\value{
a character string of original function name.
	if the function specified as fun is 'original' of it,
	this funciton returns name of fun itself.
	Note that this function cannot work correctly if a copy of the function
	and the original function are in the same environment.
}
\description{
(Internal) Find original name of the function.
}
\examples{
a <- ls
model.adapter:::find.original.name(a)

a <- glm
model.adapter:::find.original.name(a)

a <- function() print("test")
model.adapter:::find.original.name(a)

# But in the following situation, this can't work correctly.
test <- function() print("test")
a <- test
model.adapter:::find.original.name(a)
}
