% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get.class.name.r
\name{get.class.name}
\alias{get.class.name}
\title{(Internal) Find class name of a model object from function name}
\usage{
get.class.name(function.name, package.name)
}
\arguments{
\item{function.name}{a character literal of function name.}

\item{package.name}{a character literal of package name.}
}
\value{
a character literal of class name
}
\description{
If specified \code{function.name} is not supported, this function assumes
that class name is same as the \code{function.name}.
}
\examples{
model.adapter:::get.class.name("cforest")
}
