% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.r
\name{get.function}
\alias{get.function}
\title{(Internal) Get function and function name from call}
\usage{
get.function(call, type = c("function", "character"), envir = parent.frame())
}
\arguments{
\item{call}{a call from which extract function.}

\item{type}{a character vector denoting type of return value.
"function" or "character" can be used. "character" is default.}

\item{envir}{an environment where to evaluate the call.}
}
\value{
If type is "function", function object.
     If type is "character", character literal of function name.
     Note that if the function in the call is copy of other function,
     this funciton tries to find original name of it.
}
\description{
(Internal) Get function and function name from call
}
\examples{
     model.adapter:::get.function(
         substitute(glm(Sepal.Length ~ ., data = iris))
     )
}
