% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.r
\name{is.formula}
\alias{is.formula}
\title{(Internal) Check an object is formula.}
\usage{
is.formula(x)
}
\arguments{
\item{x}{an object.}
}
\value{
returns TRUE if \emph{x} is formula otherwise returns FALSE.
}
\description{
(Internal) Check an object is formula.
}
