% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.r
\name{is.s3.generic}
\alias{is.s3.generic}
\alias{is.s4.generic}
\alias{is.generic}
\title{(Internal) Is a function generic?}
\usage{
is.s3.generic(fun.name)

is.s4.generic(fun.name, package = "")

is.generic(fun.name, package = "")
}
\arguments{
\item{fun.name}{a character string naming the function.}

\item{package}{a character string of package name.}
}
\value{
returns TRUE if fun.name is generic function and FALSE otherwise.
}
\description{
This function \emph{roughly} test a function is generic.
}
\section{Functions}{
\itemize{
\item \code{is.s3.generic}: (Internal) function for S3 generic function.

\item \code{is.s4.generic}: (Internal) function for S4 generic function.
}}

\examples{
model.adapter:::is.s3.generic("plot")
model.adapter:::is.s3.generic("glm")
model.adapter:::is.s4.generic("plot")
model.adapter:::is.s4.generic("lmer")
model.adapter:::is.generic("plot")
model.adapter:::is.generic("glm")
}
