% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict.types.r
\name{make.predict.types}
\alias{make.predict.types}
\title{(Internal) Prepare predict.types.}
\usage{
make.predict.types(
  predict.types,
  response = "response",
  link = "link",
  prob = "prob",
  class = "class"
)
}
\arguments{
\item{predict.types}{a character vector of predict.types.}

\item{response}{a character representing predict type of reponse scale.}

\item{link}{a character representing predict type of link scale.}

\item{prob}{a character representing predict type of class probability.}

\item{class}{a character representing predict type of class labels.}
}
\value{
a character representing predict.types.
}
\description{
Prepare a named character vector for predict.types.
}
