% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.r
\name{match.generic.call}
\alias{match.generic.call}
\title{(Internal) match.call() with handling of generic function.}
\usage{
match.generic.call(call, envir, package = "")
}
\arguments{
\item{call}{a call to be matched.}

\item{envir}{an environment to evaluate call.}

\item{package}{a character string representing package name of the funciton.}
}
\value{
matched fully-named call.
}
\description{
Some generic functions have different arguments for different classes. In
such cases, match.call() function returns wrong results. This function
finds actual generic function called for focal class and matches call with
the function. If call is not a call for non-generic function, matching is
done by match.call() function.
}
\details{
If an object has more than two classes, the first class in the result of
class() with generic function is used.
}
\examples{
     model.adapter:::match.generic.call(
         substitute(hist(1:10)), environment()
     )
}
