% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.r
\name{match.generic.call.s3}
\alias{match.generic.call.s3}
\title{(Internal) 'match.call' considering generic functions.}
\usage{
match.generic.call.s3(call, envir)
}
\arguments{
\item{call}{a call to be matched.}

\item{envir}{the environment where calls are evaludated.}
}
\description{
(Internal) 'match.call' considering generic functions.
}
