% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model.adapter.prediction.r
\name{model.adapter.prediction-class}
\alias{model.adapter.prediction-class}
\alias{model.adapter.prediction.class}
\title{model.adapter.prediction-class}
\description{
model.adapter.prediction-class

model.adapter.prediction-class
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{fit}}{a matrix containing predicted result. For continuous variables,
     predicted values are placed in "fit" column. If an interval
     (e.g., confidence, prediction, etc.) is available, iupper and lower
     values are placed in "upper" and "lower" columns.

     For discrete variables, content of this field depends on value of
     \emph{type} field. If \emph{type} is "probability", fit is a matrix
     with columns representing probability of each class.
     If response variable is binary (0/1) variable, fit is a matrix with the
     first column representing probability of "0" and the second column
     representing probability of "1".

     For classification models and type = "class", predicted classes.}

\item{\code{type}}{a character literal representing type of prediction, one of
"response", "link", "prob" and  "class".}

\item{\code{fixed}}{a data.frame containing data used for prediction.
If original data.frame is not available, this field is NULL.}

\item{\code{interval.type}}{a character literal representing type and level of the interval.
If no interval is available, this field is NULL.
If the interval is confidence interval, this field has "confidence".
If the interval is prediction interval, this field has "prediction".}

\item{\code{interval.level}}{a number representing level of interval (0 <= interval.level <= 1).
If no interval is calculated, this field should be NULL.}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-model.adapter.prediction-new}{\code{model.adapter.prediction.class$new()}}
\item \href{#method-model.adapter.prediction-clone}{\code{model.adapter.prediction.class$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-model.adapter.prediction-new"></a>}}
\if{latex}{\out{\hypertarget{method-model.adapter.prediction-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{model.adapter.prediction.class$new(
  fit,
  type = c("response", "link", "prob", "class"),
  fixed = NULL,
  interval.type = NULL,
  interval.level = NULL,
  logical.response = FALSE
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-model.adapter.prediction-clone"></a>}}
\if{latex}{\out{\hypertarget{method-model.adapter.prediction-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{model.adapter.prediction.class$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
