% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model.adapter.prediction.r
\name{model.adapter.prediction}
\alias{model.adapter.prediction}
\title{(Internal) Initialize model.adapter.prediction object}
\usage{
model.adapter.prediction(
  fit,
  type = c("response", "link", "prob", "class"),
  fixed = NULL,
  interval.type = NULL,
  interval.level = NULL,
  logical.response = FALSE
)
}
\arguments{
\item{fit}{a matrix containing predicted result.
    For continuous response variables, predicted values should be placed in
    the first column. Also, if the prediction has interval (e.g., confidence
    interval, prediction interval, etc.) upper and lower values of them
    should be placed in the second and third columns, respectively.

    For binary model, matrix with one column with probability of positive
    case or two columns with probability of negative and positive cases.
    For classification model with >3 classes, probability of each class.

    For classification models and type = "class", predicted classes.}

\item{type}{a character literal representing type of prediction. Currently,
"response", "link", "prob" and  "class" are supported.}

\item{fixed}{a data.frame containing data used for prediction.}

\item{interval.type}{a character literal representing type of interval.}

\item{interval.level}{a numeric value representing level of interval.}

\item{logical.response}{a logical indicationg response variable is logical.}
}
\value{
an object of
	\code{\link[=model.adapter.prediction.class]{model.adapter.prediction}}
	class.
}
\description{
model.adapter.prediction function initialize model.adapter.prediction
object.
}
