% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model.interface.gam.mgcv.r
\name{model.interface.gam-class (mgcv)}
\alias{model.interface.gam-class (mgcv)}
\alias{model.interface.gam.class}
\title{(Internal) model.interface class for gam}
\description{
This reference class contains methods for \code{\link[mgcv]{gam}} in
\emph{mgcv} package.
}
\section{Super class}{
\code{\link[model.adapter:model.interface]{model.adapter::model.interface}} -> \code{model.interface.gam}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-model.interface.gam-expand.formula}{\code{model.interface.gam.class$expand.formula()}}
\item \href{#method-model.interface.gam-adjust.offset}{\code{model.interface.gam.class$adjust.offset()}}
\item \href{#method-model.interface.gam-clone}{\code{model.interface.gam.class$clone()}}
}
}
\if{html}{\out{
<details><summary>Inherited methods</summary>
<ul>
<li><span class="pkg-link" data-pkg="model.adapter" data-topic="model.interface" data-id="get.call"><a href='../../model.adapter/html/model.interface.html#method-model.interface-get.call'><code>model.adapter::model.interface$get.call()</code></a></span></li>
<li><span class="pkg-link" data-pkg="model.adapter" data-topic="model.interface" data-id="get.data"><a href='../../model.adapter/html/model.interface.html#method-model.interface-get.data'><code>model.adapter::model.interface$get.data()</code></a></span></li>
<li><span class="pkg-link" data-pkg="model.adapter" data-topic="model.interface" data-id="get.family"><a href='../../model.adapter/html/model.interface.html#method-model.interface-get.family'><code>model.adapter::model.interface$get.family()</code></a></span></li>
<li><span class="pkg-link" data-pkg="model.adapter" data-topic="model.interface" data-id="get.formula"><a href='../../model.adapter/html/model.interface.html#method-model.interface-get.formula'><code>model.adapter::model.interface$get.formula()</code></a></span></li>
<li><span class="pkg-link" data-pkg="model.adapter" data-topic="model.interface" data-id="get.link"><a href='../../model.adapter/html/model.interface.html#method-model.interface-get.link'><code>model.adapter::model.interface$get.link()</code></a></span></li>
<li><span class="pkg-link" data-pkg="model.adapter" data-topic="model.interface" data-id="get.linkinv"><a href='../../model.adapter/html/model.interface.html#method-model.interface-get.linkinv'><code>model.adapter::model.interface$get.linkinv()</code></a></span></li>
<li><span class="pkg-link" data-pkg="model.adapter" data-topic="model.interface" data-id="get.model.type"><a href='../../model.adapter/html/model.interface.html#method-model.interface-get.model.type'><code>model.adapter::model.interface$get.model.type()</code></a></span></li>
<li><span class="pkg-link" data-pkg="model.adapter" data-topic="model.interface" data-id="get.offset.names"><a href='../../model.adapter/html/model.interface.html#method-model.interface-get.offset.names'><code>model.adapter::model.interface$get.offset.names()</code></a></span></li>
<li><span class="pkg-link" data-pkg="model.adapter" data-topic="model.interface" data-id="initialize"><a href='../../model.adapter/html/model.interface.html#method-model.interface-initialize'><code>model.adapter::model.interface$initialize()</code></a></span></li>
<li><span class="pkg-link" data-pkg="model.adapter" data-topic="model.interface" data-id="predict"><a href='../../model.adapter/html/model.interface.html#method-model.interface-predict'><code>model.adapter::model.interface$predict()</code></a></span></li>
</ul>
</details>
}}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-model.interface.gam-expand.formula"></a>}}
\if{latex}{\out{\hypertarget{method-model.interface.gam-expand.formula}{}}}
\subsection{Method \code{expand.formula()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{model.interface.gam.class$expand.formula(
  f,
  d,
  specials = NULL,
  package = "mgcv"
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-model.interface.gam-adjust.offset"></a>}}
\if{latex}{\out{\hypertarget{method-model.interface.gam-adjust.offset}{}}}
\subsection{Method \code{adjust.offset()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{model.interface.gam.class$adjust.offset(
  x,
  envir,
  package,
  pred,
  newdata,
  type,
  ...
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-model.interface.gam-clone"></a>}}
\if{latex}{\out{\hypertarget{method-model.interface.gam-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{model.interface.gam.class$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
