% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model.interface.gamm.r
\name{model.interface.gamm-class (mgcv)}
\alias{model.interface.gamm-class (mgcv)}
\alias{model.interface.gamm.class}
\title{(Internal) model.interface class for gamm}
\description{
This reference class contains methods for \code{\link[mgcv]{gamm}} in
\emph{mgcv} package.
Note that because an object of gamm does not keep original call,
get.call() function always returns NULL.
}
\section{Super class}{
\code{\link[model.adapter:model.interface]{model.adapter::model.interface}} -> \code{model.interface.gamm}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-get.call}{\code{model.interface.gamm.class$get.call()}}
\item \href{#method-get.family}{\code{model.interface.gamm.class$get.family()}}
\item \href{#method-get.formula}{\code{model.interface.gamm.class$get.formula()}}
\item \href{#method-get.data}{\code{model.interface.gamm.class$get.data()}}
\item \href{#method-expand.formula}{\code{model.interface.gamm.class$expand.formula()}}
\item \href{#method-predict}{\code{model.interface.gamm.class$predict()}}
\item \href{#method-adjust.offset}{\code{model.interface.gamm.class$adjust.offset()}}
\item \href{#method-clone}{\code{model.interface.gamm.class$clone()}}
}
}
\if{html}{
\out{<details open ><summary>Inherited methods</summary>}
\itemize{
\item \out{<span class="pkg-link" data-pkg="model.adapter" data-topic="model.interface" data-id="get.link">}\href{../../model.adapter/html/model.interface.html#method-get.link}{\code{model.adapter::model.interface$get.link()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="model.adapter" data-topic="model.interface" data-id="get.linkinv">}\href{../../model.adapter/html/model.interface.html#method-get.linkinv}{\code{model.adapter::model.interface$get.linkinv()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="model.adapter" data-topic="model.interface" data-id="get.model.type">}\href{../../model.adapter/html/model.interface.html#method-get.model.type}{\code{model.adapter::model.interface$get.model.type()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="model.adapter" data-topic="model.interface" data-id="get.offset.names">}\href{../../model.adapter/html/model.interface.html#method-get.offset.names}{\code{model.adapter::model.interface$get.offset.names()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="model.adapter" data-topic="model.interface" data-id="initialize">}\href{../../model.adapter/html/model.interface.html#method-initialize}{\code{model.adapter::model.interface$initialize()}}\out{</span>}
}
\out{</details>}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-get.call"></a>}}
\if{latex}{\out{\hypertarget{method-get.call}{}}}
\subsection{Method \code{get.call()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{model.interface.gamm.class$get.call(x)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-get.family"></a>}}
\if{latex}{\out{\hypertarget{method-get.family}{}}}
\subsection{Method \code{get.family()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{model.interface.gamm.class$get.family(
  x,
  type = c("character", "family"),
  envir
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-get.formula"></a>}}
\if{latex}{\out{\hypertarget{method-get.formula}{}}}
\subsection{Method \code{get.formula()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{model.interface.gamm.class$get.formula(x, envir, package = "")}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-get.data"></a>}}
\if{latex}{\out{\hypertarget{method-get.data}{}}}
\subsection{Method \code{get.data()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{model.interface.gamm.class$get.data(x, envir, package = "", ...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-expand.formula"></a>}}
\if{latex}{\out{\hypertarget{method-expand.formula}{}}}
\subsection{Method \code{expand.formula()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{model.interface.gamm.class$expand.formula(
  f,
  d,
  specials = c("s", "te", "ti", "t2"),
  package = "mgcv"
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-predict"></a>}}
\if{latex}{\out{\hypertarget{method-predict}{}}}
\subsection{Method \code{predict()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{model.interface.gamm.class$predict(object, newdata = NULL, type, ...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-adjust.offset"></a>}}
\if{latex}{\out{\hypertarget{method-adjust.offset}{}}}
\subsection{Method \code{adjust.offset()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{model.interface.gamm.class$adjust.offset(x, envir, package, pred, ...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{model.interface.gamm.class$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
