% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model.interface.gbm.r
\name{model.interface.gbm-class (gbm)}
\alias{model.interface.gbm-class (gbm)}
\alias{model.interface.gbm.class}
\title{(Internal) model.interface class for gbm}
\description{
This reference class contains methods for \code{\link[gbm]{gbm}} in
\emph{gbm} package.
}
\section{Super class}{
\code{\link[model.adapter:model.interface]{model.adapter::model.interface}} -> \code{model.interface.gbm}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-model.interface.gbm-get.data}{\code{model.interface.gbm.class$get.data()}}
\item \href{#method-model.interface.gbm-predict}{\code{model.interface.gbm.class$predict()}}
\item \href{#method-model.interface.gbm-adjust.offset}{\code{model.interface.gbm.class$adjust.offset()}}
\item \href{#method-model.interface.gbm-get.model.type}{\code{model.interface.gbm.class$get.model.type()}}
\item \href{#method-model.interface.gbm-clone}{\code{model.interface.gbm.class$clone()}}
}
}
\if{html}{\out{
<details><summary>Inherited methods</summary>
<ul>
<li><span class="pkg-link" data-pkg="model.adapter" data-topic="model.interface" data-id="expand.formula"><a href='../../model.adapter/html/model.interface.html#method-model.interface-expand.formula'><code>model.adapter::model.interface$expand.formula()</code></a></span></li>
<li><span class="pkg-link" data-pkg="model.adapter" data-topic="model.interface" data-id="get.call"><a href='../../model.adapter/html/model.interface.html#method-model.interface-get.call'><code>model.adapter::model.interface$get.call()</code></a></span></li>
<li><span class="pkg-link" data-pkg="model.adapter" data-topic="model.interface" data-id="get.family"><a href='../../model.adapter/html/model.interface.html#method-model.interface-get.family'><code>model.adapter::model.interface$get.family()</code></a></span></li>
<li><span class="pkg-link" data-pkg="model.adapter" data-topic="model.interface" data-id="get.formula"><a href='../../model.adapter/html/model.interface.html#method-model.interface-get.formula'><code>model.adapter::model.interface$get.formula()</code></a></span></li>
<li><span class="pkg-link" data-pkg="model.adapter" data-topic="model.interface" data-id="get.link"><a href='../../model.adapter/html/model.interface.html#method-model.interface-get.link'><code>model.adapter::model.interface$get.link()</code></a></span></li>
<li><span class="pkg-link" data-pkg="model.adapter" data-topic="model.interface" data-id="get.linkinv"><a href='../../model.adapter/html/model.interface.html#method-model.interface-get.linkinv'><code>model.adapter::model.interface$get.linkinv()</code></a></span></li>
<li><span class="pkg-link" data-pkg="model.adapter" data-topic="model.interface" data-id="get.offset.names"><a href='../../model.adapter/html/model.interface.html#method-model.interface-get.offset.names'><code>model.adapter::model.interface$get.offset.names()</code></a></span></li>
<li><span class="pkg-link" data-pkg="model.adapter" data-topic="model.interface" data-id="initialize"><a href='../../model.adapter/html/model.interface.html#method-model.interface-initialize'><code>model.adapter::model.interface$initialize()</code></a></span></li>
</ul>
</details>
}}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-model.interface.gbm-get.data"></a>}}
\if{latex}{\out{\hypertarget{method-model.interface.gbm-get.data}{}}}
\subsection{Method \code{get.data()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{model.interface.gbm.class$get.data(x, envir, package = "")}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-model.interface.gbm-predict"></a>}}
\if{latex}{\out{\hypertarget{method-model.interface.gbm-predict}{}}}
\subsection{Method \code{predict()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{model.interface.gbm.class$predict(object, newdata, type, random, ...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-model.interface.gbm-adjust.offset"></a>}}
\if{latex}{\out{\hypertarget{method-model.interface.gbm-adjust.offset}{}}}
\subsection{Method \code{adjust.offset()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{model.interface.gbm.class$adjust.offset(
  x,
  envir,
  package,
  pred,
  newdata,
  type,
  ...
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-model.interface.gbm-get.model.type"></a>}}
\if{latex}{\out{\hypertarget{method-model.interface.gbm-get.model.type}{}}}
\subsection{Method \code{get.model.type()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{model.interface.gbm.class$get.model.type(x, envir, package = "", ...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-model.interface.gbm-clone"></a>}}
\if{latex}{\out{\hypertarget{method-model.interface.gbm-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{model.interface.gbm.class$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
