% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model.interface.glmer.r
\name{model.interface.glmerMod-class (lme4)}
\alias{model.interface.glmerMod-class (lme4)}
\alias{model.interface.glmerMod.class}
\title{(Internal) model.interface class for glmer}
\description{
This reference class contains methods for \code{\link[lme4]{glmer}} in
\emph{lme4} package.
}
\section{Super class}{
\code{\link[model.adapter:model.interface]{model.adapter::model.interface}} -> \code{model.interface.glmerMod}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-get.data}{\code{model.interface.glmerMod.class$get.data()}}
\item \href{#method-adjust.offset}{\code{model.interface.glmerMod.class$adjust.offset()}}
\item \href{#method-clone}{\code{model.interface.glmerMod.class$clone()}}
}
}
\if{html}{
\out{<details ><summary>Inherited methods</summary>}
\itemize{
\item \out{<span class="pkg-link" data-pkg="model.adapter" data-topic="model.interface" data-id="expand.formula">}\href{../../model.adapter/html/model.interface.html#method-expand.formula}{\code{model.adapter::model.interface$expand.formula()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="model.adapter" data-topic="model.interface" data-id="get.call">}\href{../../model.adapter/html/model.interface.html#method-get.call}{\code{model.adapter::model.interface$get.call()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="model.adapter" data-topic="model.interface" data-id="get.family">}\href{../../model.adapter/html/model.interface.html#method-get.family}{\code{model.adapter::model.interface$get.family()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="model.adapter" data-topic="model.interface" data-id="get.formula">}\href{../../model.adapter/html/model.interface.html#method-get.formula}{\code{model.adapter::model.interface$get.formula()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="model.adapter" data-topic="model.interface" data-id="get.link">}\href{../../model.adapter/html/model.interface.html#method-get.link}{\code{model.adapter::model.interface$get.link()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="model.adapter" data-topic="model.interface" data-id="get.linkinv">}\href{../../model.adapter/html/model.interface.html#method-get.linkinv}{\code{model.adapter::model.interface$get.linkinv()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="model.adapter" data-topic="model.interface" data-id="get.model.type">}\href{../../model.adapter/html/model.interface.html#method-get.model.type}{\code{model.adapter::model.interface$get.model.type()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="model.adapter" data-topic="model.interface" data-id="get.offset.names">}\href{../../model.adapter/html/model.interface.html#method-get.offset.names}{\code{model.adapter::model.interface$get.offset.names()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="model.adapter" data-topic="model.interface" data-id="initialize">}\href{../../model.adapter/html/model.interface.html#method-initialize}{\code{model.adapter::model.interface$initialize()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="model.adapter" data-topic="model.interface" data-id="predict">}\href{../../model.adapter/html/model.interface.html#method-predict}{\code{model.adapter::model.interface$predict()}}\out{</span>}
}
\out{</details>}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-get.data"></a>}}
\if{latex}{\out{\hypertarget{method-get.data}{}}}
\subsection{Method \code{get.data()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{model.interface.glmerMod.class$get.data(x, envir, package = "", ...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-adjust.offset"></a>}}
\if{latex}{\out{\hypertarget{method-adjust.offset}{}}}
\subsection{Method \code{adjust.offset()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{model.interface.glmerMod.class$adjust.offset(x, envir, package, pred, ...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{model.interface.glmerMod.class$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
