% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model.interface.MCMCglmm.r
\name{model.interface.MCMCglmm-class (MCMCglmm)}
\alias{model.interface.MCMCglmm-class (MCMCglmm)}
\alias{model.interface.MCMCglmm.class}
\title{(Internal) model.interface class for MCMCglmm}
\description{
This reference class contains methods for \code{\link[MCMCglmm]{MCMCglmm}}
in \emph{MCMCglmm} package. Because MCMCglmm object does not have original
data, this class can't obtain data from MCMCglmm objects. Also, the objects
does not have original call, this class cannot obtain call from the
objects. Link and inverse link functions are implimented only for
gaussian, poisson, categorical, multinomial, geometric and exponential
families.
}
\section{Super class}{
\code{\link[model.adapter:model.interface]{model.adapter::model.interface}} -> \code{model.interface.MCMCglmm}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-model.interface.MCMCglmm-get.data}{\code{model.interface.MCMCglmm.class$get.data()}}
\item \href{#method-model.interface.MCMCglmm-get.formula}{\code{model.interface.MCMCglmm.class$get.formula()}}
\item \href{#method-model.interface.MCMCglmm-get.family}{\code{model.interface.MCMCglmm.class$get.family()}}
\item \href{#method-model.interface.MCMCglmm-get.call}{\code{model.interface.MCMCglmm.class$get.call()}}
\item \href{#method-model.interface.MCMCglmm-get.link}{\code{model.interface.MCMCglmm.class$get.link()}}
\item \href{#method-model.interface.MCMCglmm-get.linkinv}{\code{model.interface.MCMCglmm.class$get.linkinv()}}
\item \href{#method-model.interface.MCMCglmm-get.model.type}{\code{model.interface.MCMCglmm.class$get.model.type()}}
\item \href{#method-model.interface.MCMCglmm-clone}{\code{model.interface.MCMCglmm.class$clone()}}
}
}
\if{html}{\out{
<details open><summary>Inherited methods</summary>
<ul>
<li><span class="pkg-link" data-pkg="model.adapter" data-topic="model.interface" data-id="adjust.offset"><a href='../../model.adapter/html/model.interface.html#method-model.interface-adjust.offset'><code>model.adapter::model.interface$adjust.offset()</code></a></span></li>
<li><span class="pkg-link" data-pkg="model.adapter" data-topic="model.interface" data-id="expand.formula"><a href='../../model.adapter/html/model.interface.html#method-model.interface-expand.formula'><code>model.adapter::model.interface$expand.formula()</code></a></span></li>
<li><span class="pkg-link" data-pkg="model.adapter" data-topic="model.interface" data-id="get.offset.names"><a href='../../model.adapter/html/model.interface.html#method-model.interface-get.offset.names'><code>model.adapter::model.interface$get.offset.names()</code></a></span></li>
<li><span class="pkg-link" data-pkg="model.adapter" data-topic="model.interface" data-id="initialize"><a href='../../model.adapter/html/model.interface.html#method-model.interface-initialize'><code>model.adapter::model.interface$initialize()</code></a></span></li>
<li><span class="pkg-link" data-pkg="model.adapter" data-topic="model.interface" data-id="predict"><a href='../../model.adapter/html/model.interface.html#method-model.interface-predict'><code>model.adapter::model.interface$predict()</code></a></span></li>
</ul>
</details>
}}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-model.interface.MCMCglmm-get.data"></a>}}
\if{latex}{\out{\hypertarget{method-model.interface.MCMCglmm-get.data}{}}}
\subsection{Method \code{get.data()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{model.interface.MCMCglmm.class$get.data(x, envir, package = "", ...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-model.interface.MCMCglmm-get.formula"></a>}}
\if{latex}{\out{\hypertarget{method-model.interface.MCMCglmm-get.formula}{}}}
\subsection{Method \code{get.formula()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{model.interface.MCMCglmm.class$get.formula(x, envir, package = "")}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-model.interface.MCMCglmm-get.family"></a>}}
\if{latex}{\out{\hypertarget{method-model.interface.MCMCglmm-get.family}{}}}
\subsection{Method \code{get.family()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{model.interface.MCMCglmm.class$get.family(
  x,
  type = c("character", "family"),
  envir
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-model.interface.MCMCglmm-get.call"></a>}}
\if{latex}{\out{\hypertarget{method-model.interface.MCMCglmm-get.call}{}}}
\subsection{Method \code{get.call()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{model.interface.MCMCglmm.class$get.call(x)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-model.interface.MCMCglmm-get.link"></a>}}
\if{latex}{\out{\hypertarget{method-model.interface.MCMCglmm-get.link}{}}}
\subsection{Method \code{get.link()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{model.interface.MCMCglmm.class$get.link(x, envir)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-model.interface.MCMCglmm-get.linkinv"></a>}}
\if{latex}{\out{\hypertarget{method-model.interface.MCMCglmm-get.linkinv}{}}}
\subsection{Method \code{get.linkinv()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{model.interface.MCMCglmm.class$get.linkinv(x, envir)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-model.interface.MCMCglmm-get.model.type"></a>}}
\if{latex}{\out{\hypertarget{method-model.interface.MCMCglmm-get.model.type}{}}}
\subsection{Method \code{get.model.type()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{model.interface.MCMCglmm.class$get.model.type(x, envir, package = "", ...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-model.interface.MCMCglmm-clone"></a>}}
\if{latex}{\out{\hypertarget{method-model.interface.MCMCglmm-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{model.interface.MCMCglmm.class$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
