% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model.interface.default.r,
%   R/model.interface.MCMCglmm.r, R/model.interface.averaging.r,
%   R/model.interface.cforest.r, R/model.interface.ctree.r,
%   R/model.interface.gam.gam.r, R/model.interface.gam.mgcv.r,
%   R/model.interface.gamm.r, R/model.interface.gbm.r, R/model.interface.glm.r,
%   R/model.interface.glmer.r, R/model.interface.glmmML.r,
%   R/model.interface.glmmTMB.r, R/model.interface.glmmadmb.r,
%   R/model.interface.lm.r, R/model.interface.lme.r, R/model.interface.lmer.r,
%   R/model.interface.randomForest.r, R/model.interface.ranger.r,
%   R/model.interface.rpart.r, R/model.interface.svm.r, R/model.interface.tree.r
\name{model.interface}
\alias{model.interface}
\alias{model.interface.default}
\alias{model.interface.MCMCglmm}
\alias{model.interface.averaging}
\alias{model.interface.RandomForest}
\alias{model.interface.BinaryTree}
\alias{model.interface.Gam}
\alias{model.interface.gam}
\alias{model.interface.gamm}
\alias{model.interface.gbm}
\alias{model.interface.glm}
\alias{model.interface.glmerMod}
\alias{model.interface.glmmML}
\alias{model.interface.glmmTMB}
\alias{model.interface.glmmadmb}
\alias{model.interface.lm}
\alias{model.interface.lme}
\alias{model.interface.lmerMod}
\alias{model.interface.randomForest}
\alias{model.interface.ranger}
\alias{model.interface.rpart}
\alias{model.interface.svm}
\alias{model.interface.tree}
\title{(Internal) Initialize model.interface class.}
\usage{
model.interface(x = NULL, ...)

\method{model.interface}{default}(...)

\method{model.interface}{MCMCglmm}(...)

\method{model.interface}{averaging}(...)

\method{model.interface}{RandomForest}(...)

\method{model.interface}{BinaryTree}(...)

\method{model.interface}{Gam}(...)

\method{model.interface}{gam}(...)

\method{model.interface}{gamm}(...)

\method{model.interface}{gbm}(...)

\method{model.interface}{glm}(...)

\method{model.interface}{glmerMod}(...)

\method{model.interface}{glmmML}(...)

\method{model.interface}{glmmTMB}(...)

\method{model.interface}{glmmadmb}(...)

\method{model.interface}{lm}(...)

\method{model.interface}{lme}(...)

\method{model.interface}{lmerMod}(...)

\method{model.interface}{randomForest}(...)

\method{model.interface}{ranger}(...)

\method{model.interface}{rpart}(...)

\method{model.interface}{svm}(...)

\method{model.interface}{tree}(...)
}
\arguments{
\item{x}{an object of supported models.
This is only used for dispatching S3 methods.}

\item{...}{not used.}
}
\value{
an object of derived class of \code{\link{model.interface-class}}.
}
\description{
This function makes an object of a derived class of \emph{model.interface}
class that abstracts differences in specifications of supported modeling
functions.
}
\section{Methods (by class)}{
\itemize{
\item \code{default}: default S3 method

\item \code{MCMCglmm}: S3 method for class 'MCMCglmm'

\item \code{averaging}: S3 method for class 'averaging'

\item \code{RandomForest}: S3 method for class 'RandomForest'

\item \code{BinaryTree}: S3 method for class 'BinaryTree'

\item \code{Gam}: S3 method for class 'Gam'

\item \code{gam}: S3 method for class 'gam'

\item \code{gamm}: S3 method for class 'gamm'

\item \code{gbm}: S3 method for class 'gbm'

\item \code{glm}: S3 method for class 'glm'

\item \code{glmerMod}: S3 method for class 'glmerMod'

\item \code{glmmML}: S3 method for class 'glmmML'

\item \code{glmmTMB}: S3 method for class 'glmmTMB'

\item \code{glmmadmb}: S3 method for class 'glmmadmb'

\item \code{lm}: S3 method for class 'lm'

\item \code{lme}: S3 method for class 'lme'

\item \code{lmerMod}: S3 method for class 'lmerMod'

\item \code{randomForest}: S3 method for class 'randomForest'

\item \code{ranger}: S3 method for class 'ranger'

\item \code{rpart}: S3 method for class 'rpart'

\item \code{svm}: S3 method for class 'svm'

\item \code{tree}: S3 method for class 'tree'
}}

\section{Adding support for new function}{

	To be continued...
}

