% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package.name.r
\name{package.name}
\alias{package.name}
\alias{package.name.default}
\alias{package.name.character}
\alias{package.name.call}
\title{(Internal) Get package name from function name or object.}
\usage{
package.name(x, envir)

\method{package.name}{default}(x, envir)

\method{package.name}{character}(x, envir)

\method{package.name}{call}(x, envir)
}
\arguments{
\item{x}{function name, call or model object.}

\item{envir}{an environment where \code{x} is evaluated.}
}
\value{
a character representing package name.
}
\description{
(Internal) Get package name from function name or object.
}
\section{Methods (by class)}{
\itemize{
\item \code{package.name(default)}: Default S3 method.

\item \code{package.name(character)}: Method for character.

\item \code{package.name(call)}: Method for call.

}}
\examples{
# Get package name from function name
require(party)
model.adapter:::package.name("cforest", environment())
# Get package name from function call.
model.adapter:::package.name(
     substitute(ctree(Petal.Length~., data=iris)), environment()
)
# Get package name from object.
data(iris)
object <- lm(Petal.Length ~ ., data = iris)
model.adapter:::package.name(object, environment())
}
