% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model.interface.glmmML.r
\name{predict.glmmML}
\alias{predict.glmmML}
\title{predict method for glmmML object.}
\usage{
\method{predict}{glmmML}(
  object,
  newdata,
  type = c("response", "link"),
  conditional = FALSE,
  ...
)
}
\arguments{
\item{object}{a glmmML object.}

\item{newdata}{a data.frame containing data used for prediction. If missing, the data
used for the modeling is used. However, because glmmML object does not
contain original data used for the modeling, retrieving the data used
for the modeling may fail if the data is removed from workspace.}

\item{type}{type of prediction. Default is scale of response variable.
If "link" is specified, prediction is of link scale is calculated.}

\item{conditional}{if FALSE, marginal (using fixed terms only and unconditioned to random
effect) predicted values are calculated. If TRUE, predicted values
conditioned to the random effect is calculated.}

\item{...}{further arguments passed to other methods.}
}
\description{
This is a predict method for \emph{glmmML} object.
}
