% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.r
\name{strip.function.call}
\alias{strip.function.call}
\title{(Internal) Strip function call from variable name(s)}
\usage{
strip.function.call(x, maxit = 100)
}
\arguments{
\item{x}{a vector of variable names.}

\item{maxit}{maximum number of iterations for nested functions.}
}
\description{
This function removes function call (i.e. 'function(' and ')') from list
of variable names.
}
