% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.r
\name{strip.offset.in.colnames}
\alias{strip.offset.in.colnames}
\title{(Internal) Strip offset from column names}
\usage{
strip.offset.in.colnames(df)
}
\arguments{
\item{df}{a data.frame.}
}
\description{
(Internal) Strip offset from column names
}
