% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.r
\name{x.names.from.formula}
\alias{x.names.from.formula}
\title{(Internal) Get names of explanatory variables (excluding random factor) from formula.}
\usage{
x.names.from.formula(
  formula,
  data = NULL,
  specials = NULL,
  type = c("all", "base")
)
}
\arguments{
\item{formula}{a formula object from which variable names are obtained.}

\item{data}{a data.frame used to expand . in the formula.}

\item{specials}{special characterss passed to \code{\link[stats]{terms.formula}}.}

\item{type}{a character literal specifying type of explanatory variables to get.}
}
\value{
a character vector of names of explanatory variables.
	If \emph{type} is "all", returns all explanatory variables in the
	formula as is.
	If \emph{type} is "base", this function returns all explanatory
	variables in their basic form.
}
\description{
(Internal) Get names of explanatory variables (excluding random factor) from formula.
}
\examples{
	# Getting explanatory variables from formula.
	# . in formula is expanded to full form.
	data(iris)
	f <- Sepal.Length ~ .
	model.adapter:::x.names.from.formula(f, data = iris)

	# Getting explanatory variables in their basic form.
	f <- Sepal.Length ~ Petal.Length + Petal.Length:Species + I(Sepal.Width^2)
	model.adapter:::x.names.from.formula(f, data = iris, type = "base")
}
