% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model.interface.averaging.r
\name{model.interface.averaging-class}
\alias{model.interface.averaging-class}
\alias{model.interface.averaging.class}
\title{(Internal) model.interface class for averaging}
\description{
This reference class contains methods for \emph{averaging} class in
\emph{MuMIn} package.
Current support for the \emph{averaging} is limited and may not work for
all supported models by \emph{MuMIn} package.
Also because of it nature, using \emph{averaging} with \emph{model.adapter}
have several limitations.
}
\details{
\describe{
	\item{\code{get.formula}} {
		because nature of \emph{\code{averaging}} object,
		this class constructs formula from all variable names in
		\emph{\code{sw}} field of the object.
	}
	\item{\code{get.familiy}, \code{get.data}} {
		family and data are obtained from the first model of the
		\emph{\code{model.calls}} or \emph{\code{modelList}} attributes
		depending on the object.
	}
	\item{\code{adjust.offset}} {
		not implemented.
	}
}
}
\section{Super class}{
\code{\link[model.adapter:model.interface]{model.adapter::model.interface}} -> \code{model.interface.averaging}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-get.formula}{\code{model.interface.averaging.class$get.formula()}}
\item \href{#method-get.family}{\code{model.interface.averaging.class$get.family()}}
\item \href{#method-get.data}{\code{model.interface.averaging.class$get.data()}}
\item \href{#method-get.offset.names}{\code{model.interface.averaging.class$get.offset.names()}}
\item \href{#method-clone}{\code{model.interface.averaging.class$clone()}}
}
}
\if{html}{
\out{<details ><summary>Inherited methods</summary>}
\itemize{
\item \out{<span class="pkg-link" data-pkg="model.adapter" data-topic="model.interface" data-id="adjust.offset">}\href{../../model.adapter/html/model.interface.html#method-adjust.offset}{\code{model.adapter::model.interface$adjust.offset()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="model.adapter" data-topic="model.interface" data-id="expand.formula">}\href{../../model.adapter/html/model.interface.html#method-expand.formula}{\code{model.adapter::model.interface$expand.formula()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="model.adapter" data-topic="model.interface" data-id="get.call">}\href{../../model.adapter/html/model.interface.html#method-get.call}{\code{model.adapter::model.interface$get.call()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="model.adapter" data-topic="model.interface" data-id="get.link">}\href{../../model.adapter/html/model.interface.html#method-get.link}{\code{model.adapter::model.interface$get.link()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="model.adapter" data-topic="model.interface" data-id="get.linkinv">}\href{../../model.adapter/html/model.interface.html#method-get.linkinv}{\code{model.adapter::model.interface$get.linkinv()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="model.adapter" data-topic="model.interface" data-id="get.model.type">}\href{../../model.adapter/html/model.interface.html#method-get.model.type}{\code{model.adapter::model.interface$get.model.type()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="model.adapter" data-topic="model.interface" data-id="initialize">}\href{../../model.adapter/html/model.interface.html#method-initialize}{\code{model.adapter::model.interface$initialize()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="model.adapter" data-topic="model.interface" data-id="predict">}\href{../../model.adapter/html/model.interface.html#method-predict}{\code{model.adapter::model.interface$predict()}}\out{</span>}
}
\out{</details>}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-get.formula"></a>}}
\if{latex}{\out{\hypertarget{method-get.formula}{}}}
\subsection{Method \code{get.formula()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{model.interface.averaging.class$get.formula(x, envir, package = "")}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-get.family"></a>}}
\if{latex}{\out{\hypertarget{method-get.family}{}}}
\subsection{Method \code{get.family()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{model.interface.averaging.class$get.family(
  x,
  type = c("character", "family"),
  envir
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-get.data"></a>}}
\if{latex}{\out{\hypertarget{method-get.data}{}}}
\subsection{Method \code{get.data()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{model.interface.averaging.class$get.data(x, envir, package = "", ...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-get.offset.names"></a>}}
\if{latex}{\out{\hypertarget{method-get.offset.names}{}}}
\subsection{Method \code{get.offset.names()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{model.interface.averaging.class$get.offset.names(x, envir, package = "")}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{model.interface.averaging.class$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
