% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model.interface.gamm.r
\name{model.interface.gamm-class (mgcv)}
\alias{model.interface.gamm-class (mgcv)}
\alias{model.interface.gamm.class}
\title{(Internal) model.interface class for gamm}
\description{
This reference class contains methods for \code{\link[mgcv]{gamm}} in
\emph{mgcv} package.
Note that because an object of gamm does not keep original call,
get.call() function always returns NULL.
}
\section{Super class}{
\code{\link[model.adapter:model.interface]{model.adapter::model.interface}} -> \code{model.interface.gamm}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-model.interface.gamm-get.call}{\code{model.interface.gamm.class$get.call()}}
\item \href{#method-model.interface.gamm-get.family}{\code{model.interface.gamm.class$get.family()}}
\item \href{#method-model.interface.gamm-get.formula}{\code{model.interface.gamm.class$get.formula()}}
\item \href{#method-model.interface.gamm-get.data}{\code{model.interface.gamm.class$get.data()}}
\item \href{#method-model.interface.gamm-expand.formula}{\code{model.interface.gamm.class$expand.formula()}}
\item \href{#method-model.interface.gamm-predict}{\code{model.interface.gamm.class$predict()}}
\item \href{#method-model.interface.gamm-adjust.offset}{\code{model.interface.gamm.class$adjust.offset()}}
\item \href{#method-model.interface.gamm-clone}{\code{model.interface.gamm.class$clone()}}
}
}
\if{html}{\out{
<details open><summary>Inherited methods</summary>
<ul>
<li><span class="pkg-link" data-pkg="model.adapter" data-topic="model.interface" data-id="get.link"><a href='../../model.adapter/html/model.interface.html#method-model.interface-get.link'><code>model.adapter::model.interface$get.link()</code></a></span></li>
<li><span class="pkg-link" data-pkg="model.adapter" data-topic="model.interface" data-id="get.linkinv"><a href='../../model.adapter/html/model.interface.html#method-model.interface-get.linkinv'><code>model.adapter::model.interface$get.linkinv()</code></a></span></li>
<li><span class="pkg-link" data-pkg="model.adapter" data-topic="model.interface" data-id="get.model.type"><a href='../../model.adapter/html/model.interface.html#method-model.interface-get.model.type'><code>model.adapter::model.interface$get.model.type()</code></a></span></li>
<li><span class="pkg-link" data-pkg="model.adapter" data-topic="model.interface" data-id="get.offset.names"><a href='../../model.adapter/html/model.interface.html#method-model.interface-get.offset.names'><code>model.adapter::model.interface$get.offset.names()</code></a></span></li>
<li><span class="pkg-link" data-pkg="model.adapter" data-topic="model.interface" data-id="initialize"><a href='../../model.adapter/html/model.interface.html#method-model.interface-initialize'><code>model.adapter::model.interface$initialize()</code></a></span></li>
</ul>
</details>
}}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-model.interface.gamm-get.call"></a>}}
\if{latex}{\out{\hypertarget{method-model.interface.gamm-get.call}{}}}
\subsection{Method \code{get.call()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{model.interface.gamm.class$get.call(x)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-model.interface.gamm-get.family"></a>}}
\if{latex}{\out{\hypertarget{method-model.interface.gamm-get.family}{}}}
\subsection{Method \code{get.family()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{model.interface.gamm.class$get.family(
  x,
  type = c("character", "family"),
  envir
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-model.interface.gamm-get.formula"></a>}}
\if{latex}{\out{\hypertarget{method-model.interface.gamm-get.formula}{}}}
\subsection{Method \code{get.formula()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{model.interface.gamm.class$get.formula(x, envir, package = "")}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-model.interface.gamm-get.data"></a>}}
\if{latex}{\out{\hypertarget{method-model.interface.gamm-get.data}{}}}
\subsection{Method \code{get.data()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{model.interface.gamm.class$get.data(x, envir, package = "", ...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-model.interface.gamm-expand.formula"></a>}}
\if{latex}{\out{\hypertarget{method-model.interface.gamm-expand.formula}{}}}
\subsection{Method \code{expand.formula()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{model.interface.gamm.class$expand.formula(
  f,
  d,
  specials = c("s", "te", "ti", "t2"),
  package = "mgcv"
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-model.interface.gamm-predict"></a>}}
\if{latex}{\out{\hypertarget{method-model.interface.gamm-predict}{}}}
\subsection{Method \code{predict()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{model.interface.gamm.class$predict(object, newdata = NULL, type, ...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-model.interface.gamm-adjust.offset"></a>}}
\if{latex}{\out{\hypertarget{method-model.interface.gamm-adjust.offset}{}}}
\subsection{Method \code{adjust.offset()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{model.interface.gamm.class$adjust.offset(x, envir, package, pred, ...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-model.interface.gamm-clone"></a>}}
\if{latex}{\out{\hypertarget{method-model.interface.gamm-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{model.interface.gamm.class$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
